Renamed SearchConsumer to WithSearch
