+++ packages/react-search-ui/src/index.js
@@ -1,6 +1,6 @@
 export { default as withSearch } from "./withSearch";
 export { default as SearchContext } from "./SearchContext";
-export { default as SearchConsumer } from "./SearchConsumer";
+export { default as WithSearch } from "./WithSearchRenderProps";
 export { default as SearchProvider } from "./SearchProvider";
 
 export * from "./containers";
